+++ sdk/keyvault/provisioning-keyvault/CHANGELOG.md
@@ -5,3 +5,7 @@
 ### Features Added
 
 - Initial preview of Azure Key Vault provisioning resources generated from the `2026-02-01` TypeSpec contract. [#39825](https://github.com/Azure/azure-sdk-for-js/pull/39825)
+
+### Bugs Fixed
+
+- Restrict access-policy operation kinds to `add`, `replace`, or `remove`. [#39825](https://github.com/Azure/azure-sdk-for-js/pull/39825)